+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/busan/RoadBusanThemeHandler.java
@@ -1,6 +1,6 @@
 package com.hubo.gillajabi.crawl.domain.service.busan;
 
-import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
+import com.hubo.gillajabi.crawl.application.dto.response.RoadCrawlResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
