+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/busan/RoadCrawlBusanCourseHandler.java
@@ -1,6 +1,6 @@
 package com.hubo.gillajabi.crawl.domain.service.busan;
 
-import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
+import com.hubo.gillajabi.crawl.application.dto.response.RoadCrawlResponse;
 import org.springframework.stereotype.Service;
 
 @Service
