+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/duru/RoadCourseDuruHandler.java
@@ -1,6 +1,6 @@
 package com.hubo.gillajabi.crawl.domain.service.duru;
 
-import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
+import com.hubo.gillajabi.crawl.application.dto.response.RoadCrawlResponse;
 import com.hubo.gillajabi.crawl.domain.entity.City;
 import com.hubo.gillajabi.crawl.domain.entity.Course;
 import com.hubo.gillajabi.crawl.domain.entity.CourseDetail;
