+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/duru/RoadThemeDuruHandler.java
@@ -1,7 +1,7 @@
 package com.hubo.gillajabi.crawl.domain.service.duru;
 
 
-import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
+import com.hubo.gillajabi.crawl.application.dto.response.RoadCrawlResponse;
 import com.hubo.gillajabi.crawl.domain.entity.CourseTheme;
 import com.hubo.gillajabi.crawl.infrastructure.dto.response.ApiThemeResponse;
 import lombok.RequiredArgsConstructor;
