+++ src/test/java/com/hubo/gillajabi/crawl/domain/service/duru/RoadCourseDuruHandlerTest.java
@@ -1,6 +1,6 @@
 package com.hubo.gillajabi.crawl.domain.service.duru;
 
-import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
+import com.hubo.gillajabi.crawl.application.dto.response.RoadCrawlResponse;
 import com.hubo.gillajabi.crawl.domain.constant.CourseLevel;
 import com.hubo.gillajabi.crawl.domain.constant.Province;
 import com.hubo.gillajabi.crawl.domain.entity.*;
